Feature validate (#19)
* wip

* wip solution validation

* Update validate.py

* _validate_side_img

+++ iquail/run.py
@@ -11,6 +11,7 @@ from .builder import Builder
 from .manager import Manager
 from .controller import ControllerConsole
 from .helper import misc
+from .validate import Validate
 
 logger = logging.getLogger(__name__)
 
@@ -43,6 +44,9 @@ def parse_args():
                         action="store_true",
                         help="Tells iQuail to install a polkit authorization file in /usr/bin/polkit-1/actions and "
                              "then rerun itself with pkexec (Linux only)")
+    parser.add_argument(Constants.ARGUMENT_VALIDATE,
+                        type=str,
+                        help="Validate a solution")
 
     return parser.parse_known_args()
 
@@ -60,6 +64,10 @@ def run(solution, installer, builder=None, controller=None, conf_ignore=None):
         controller = ControllerConsole()
     manager = Manager(installer, solution, builder, controller.is_graphical(),
                       conf_ignore=conf_ignore)
+    if args.iquail_validate:
+        v = Validate(os.path.realpath(
+            args.iquail_validate), installer)
+        exit(v.run())
     controller.setup(manager)
     if args.iquail_rm:
         shutil.rmtree(args.iquail_rm)
